Halve semgrep --version runtime
...by delaying a bunch of imports and reducing import burden
